It gets worse
- Added self check in PNFBatch.migrate.
- Added docstring to PNFGroup
- Pulled out label into `core.pnf_label`
- Commented out a chunk of leftover migration code that still checked for `None`
- Debug pane now uses contexts (well duh, the other option doesn't work anymore)
