+++ pyday_night_funkin/core/pyglet_tl_patch/__init__.py
@@ -6,7 +6,6 @@ from loguru import logger
 if pyglet.version != "2.0.dev13":
 	logger.warning("TL patches may fail!")
 
-from pyday_night_funkin.core.pyglet_tl_patch.tl_label import TLLabel
 from pyday_night_funkin.core.pyglet_tl_patch.tl_shapes import TLRectangle
 
-__all__ = ["TLLabel", "TLRectangle"]
+__all__ = ["TLRectangle"]
